+++ src/main/java/net/minestom/server/entity/metadata/animal/SheepMeta.java
@@ -12,13 +12,13 @@ public class SheepMeta extends AnimalMeta {
 
     public int getColor() {
         // TODO: make COLOR_ID is 4 bits
-        byte bitMask = (byte) ((MetadataDef.Entry.Mask) MetadataDef.Sheep.COLOR_ID).bitMask();
+        byte bitMask = ((MetadataDef.Entry.Mask) MetadataDef.Sheep.COLOR_ID).bitMask();
         return metadata.get(MetadataDef.Sheep.SHEEP_FLAGS) & bitMask;
     }
 
     public void setColor(byte color) {
         // TODO: make COLOR_ID is 4 bits
-        byte bitMask = (byte) ((MetadataDef.Entry.Mask) MetadataDef.Sheep.COLOR_ID).bitMask();
+        byte bitMask = ((MetadataDef.Entry.Mask) MetadataDef.Sheep.COLOR_ID).bitMask();
         byte before = metadata.get(MetadataDef.Sheep.SHEEP_FLAGS);
         byte mask = before;
         mask &= ~(bitMask);
